requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.4
deep-translator==1.11.4
distro==1.9.0
dnspython==2.7.0
fpdf==1.7.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
groq==0.3.0
grpc-stubs==1.53.0.6
grpcio==1.73.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langdetect==1.0.9
libretranslatepy==2.1.1
lxml==5.4.0
MarkupSafe==3.0.2
multidict==6.4.4
narwhals==1.39.1
numpy==2.2.6
openai==0.28.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.25.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
streamlit==1.45.1
streamlit-webrtc==0.62.4
tenacity==9.1.2
toml==0.10.2
tornado==6.5
tqdm==4.67.1
translate==3.6.1
types-protobuf==4.25.0.20240417
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.1
